+++ common/kotlinx-coroutines-core-common/test/SupervisorTest.kt
@@ -195,6 +195,30 @@ class SupervisorTest : TestBase() {
         assertTrue(parent.isCancelled)
     }
 
+    @Test
+    fun testSupervisorScopeCancellationVsException() = runTest {
+        expect(1)
+        var job: Job? = null
+        job = launch(start = CoroutineStart.UNDISPATCHED) {
+            expect(2)
+            try {
+                supervisorScope {
+                    expect(3)
+                    yield() // must suspend
+                    expect(5)
+                    job!!.cancel() // cancel this job _before_ it throws
+                    throw TestException1()
+                }
+            } catch (e: TestException1) {
+                // must have caught TextException
+                expect(6)
+            }
+        }
+        expect(4)
+        yield() // to coroutineScope
+        finish(7)
+    }
+
     private class TestException1 : Exception()
     private class TestException2 : Exception()
 }